requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
cvxopt==1.2.3
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
Markdown==3.1.1
mock==3.0.5
numpy==1.16.3
protobuf==3.7.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
tqdm==4.32.1
Werkzeug==0.15.4
